unit test for categories & payment

+++ src/components/category/services/deleteById.js
@@ -2,6 +2,10 @@ import Category from '../category.model';
 
 const deleteById = async (id) => {
   try {
+    const category = await Category.findById(id);
+    if (!category) {
+      throw new Error('Category not found');
+    }
     await Category.findByIdAndDelete(id);
     return true;
   } catch (error) {
